feat and fix: properly implement exclude and include
The culprit for the bug was actually `walkdir`. I have learnt that it's not
a good crate for filtering paths.

As a "walker", it's very suitable for tasks for filtering (with `.filter`)
such as if a path is a file or a directory, and sorting files e.g. by names or
by key. However, the reason the bug happens is the innate nature of the walker.

Specifically, there is an issue with using the `.filter_entry` as it does
not filter correctly the set of conditions. All the while, I was banging
my head for why this happens. It turns out, the filter_entry didn't read my
`exclude_path.contains`.

In the future, I will replace `walkdir` with a handrolled implementation
of just using the standard library and Iterator traits. That's low priority
but I will probably make it happen in the future.

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>