fix: `remote_checkout_branch` creating a branch from remote branch with incorrect name
The logic is correct to get the shorthand name since if it does not have a shorthand
name, you'll get a full reference name.

- shorthand: `<remote name>/<branch name>`
- full reference: `refs/remotes/<remote name>/<branch name>`

Since the passed `branch` parameter is always a remote branch, hence, why
the function is named `remote_checkout_branch`, the `refremote_path`
variable formats the full ref path until the remote name for string
splitting in the next line. The rest of the code is then self-explanatory.

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>