improvement: remove any usage of walkdir
in this case, we are using our own implementation. the sorting as well is
based on the full path name rather than sorting by file-name. however,
resulting hashes will be different as we are not using the sort used in
walkdir. fortunately, it will still ensure the deterministic order, you just
need to remove the generated tarball using a version of roast before this commit.

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>