refactor: rework recomprizz renaming logic
with this, a user can either pass a hard-code name. However,
if they pass the `--renamepattern`, it is assumed that the
`--rename` flag contains a valid regex string to process the renaming.

if both flags are not passed, we will try to remove the old file extension
by using known extensions for supported formats, otherwise, we will just
use the `.file_name()` method to just get the old file name. this allows
some flexibility in case a file is supported with a known mime-type but
has an unexpected file extension e.g. instead of `.tar.gz` for gzipped tarballs,
the user accidentally had named it `.tar.zst` but the mime-type is actually
a gzipped tarball so the rename will result to `nameoffile.tar.zst.tar.gz` since
they passed a compression `gz`.

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>